Dynamic programming optimization, coins
Succeeded in optimizing it, the time complexity changed from 3 square to 2square. The experiece is so painful. Must be clear with what you are doing, make sure you know every detail before coding. Otherwise the code you product will create wrong answer in the OJ system. This time I tried to code but I didn't know the key method to optimize the code. So i took almost 4 hours in vain. At last, I came up with the method, and used only 10 minites to finish the code.