DP， Map, LSC, DFS
The prior way must be seleted when using dfs algorithm. A return statement is also suggested. The previous RE code makes matter in 
if (map[xx][y] == map[x][y]) {//在这里X是向上走
            BFS(map, xx, y, str2, flag);
            return;
        }
        if (map[x][yy] == map[x][yy])
            BFS(map, x, yy, str2, flag);
The order of the two if sentences determines the priority of them and the return sentence will end the recursive.