Priorityqueue, Greedy, pre-process ,Scan twice, NI
The difficulty of this problem is to maintain the leftside sum. I came up with the method to deal with the rightside but i got no idea about how to deat with the leftside. Use the foolish method that generate sum from index C - N/2 to 0 will surely TLE.
The right way to deal with the leftside elements is to do the same thing as the rightside, scan it from index 0 to C - N/2 instead of scanning from C - N/2 to 0.
I had seen the thinking of scan twice in dp problem, yet i dont have a good command of it...
Do better next time! That's an inspired problem, if you know the method to deal with the element in one side, why don't you do the same thing to the other side?