BIT, NI
Classic implement of bit.
Sort the cows according to their volume. So for the i th cow, the cow from 1 to i - 1 need to use the same volume as i th cow.
Suppose the ith cow's pos is x, lcnt cows is on the left side of cow i, i - lcnt cows is on the right side of cow i. 
Let lsum is sum of distance of cow on the left side of cow i. Thus totaldis - lsum is the right cow's distance.
The volume used to communicate with cow i become vol * ((lcnt) * x -  lsum) +  vol * (td - lsum - (i - lcnt) * x)

use two bit to maintain it.
*Noticeable, update is carried after calculating ans.
* In add the function, the terminate condition is not i < N but i < MAXP(Here MAXP is equal to MAXN).
*The initial value of ll in struct array is 0.